Do not try to define libvirt domain if it isn't starting
Libvirt domain object, even if defined, isn't guaranteed to be up to
date, if the domain isn't running. Do not pretend it is. QubesVM.start()
method explicitly define the domain in libvirt. When accessing
QubesVM.libvirt_domain property, return None if it isn't defined already
(instead of trying to define it) - as the docstring already says.

The immediate reason is that accessing Volume.block_device() isn't
guaranteed to work before Volume.start(). But this change also speedups
getting -1 XID if VM is not running - it definitely isn't, if it isn't
defined in libvirt and there is no sense in defining it just to confirm
this.

While at it, remove redundant checking for QubesVM.libvirt_domain being
none - QubesVM.is_running() already handles this case. And update the
tests to take this into account.