qubesd: at startup, stop storage for stopped qubes
If qubesd is not cleanly shut down, it is possible that storage was also
not properly stopped.  Therefore, qubesd needs to stop storage for all
domains that are not currently running.  Otherwise, it is possible that
changes to the private volumes of disposable qubes will persist when
they should not, for example.

Stopping storage can be a slow process, due to limitations in the
dm-thin kernel driver.  To ensure the startup timeout doesn’t expire,
use an EXTEND_TIMEOUT_USEC= message to extend it to now + 60 seconds
every 30 seconds.  The “STATUS=Cleaning up storage for stopped qubes”
part of the message tells systemd that qubesd is cleaning up storage.
the human-readable output of various tools, such as “systemctl status”.
Once qubesd has stopped all the stale storage and has fully started, the
status is changed to “qubesd online and operational”.