storage/file: support online resize of save_on_stop=True volumes
After resizing a persistent volume of a running VM, not only loop device
needs to be resized, but also device-mapper (snapshot-origin) built on
top of it. Otherwise VM won't see the new size immediately.
Note the CoW image file doesn't need resize, because it is about keeping
differences compared to the VM start time - at which point it had still
the old size.