_kill_locked(): Wait for VIR_DOMAIN_EVENT_STOPPED
This ensures that the domain is actually dead, and that a spurious
stopped event will not happen.  Furthermore, it ensures that the storage
has been released, even if the libvirt API call returns immediately
rather than waiting for the VM to actually be destroyed.

Prior to encrypting volatile volumes, this bug did not actually cause
problems.  For the file and reflink pools, the underlying loop device is
clenaed up by the block script, and the file on disk is deleted when the
last reference to it goes away.  The LVM2 pool, on the other hand, calls
‘blkdiscard’ before removing the storage.  ‘blkdiscard’ does _not_
require the block device to be unused to have stopped, and is slow on
thin pools due to a Linux kernel bug.  Therefore, by the time LVM2 is
actually called to removed the volume, the VM will likely have actually
stopped, meaning that the volume will no longer be in use.

However, ‘cryptsetup close’ is called _before_ ‘blkdiscard’, and it
_does_ require the volume to be unused.  This makes the problem much
more likely to show up.  ‘cryptsetup close’ has built-in retry logic, so
if the VM stops soon enough, the only symptom will be a ‘device-mapper:
remove ioctl on <device name> failed: device or resource busy’ warning
printed to stderr.  However, if the VM takes too long to stop,
‘cryptsetup close’ will fail outright, preventing the storage from
stopping.

Furthermore, if ‘VIR_DOMAIN_EVENT_STOPPED’ is delivered at exactly the
wrong time, it is possible for the storage to be stopped _twice_.  This
results in bad behavior, such as LVM2 failing with ENOENT.