tests/dispvm: simplify editor handling
Instead of separately detecting editor window and interacting with it,
call _handle_editor() directly on any window found. If it complains it
doesn't know this editor (by raising KeyError), wait for another window
and retry - up to 5 times. This avoids duplicated logic, and should be
more resiliant for splash windows etc.