storage/reflink: volume usage(): omit wrong docstring
It's neither correct to say that this is "volume disk usage from the
VM's perspective" (because an image file can be sparsified on the host,
e.g. by restoring a backup, cloning a VM, or 'fallocate --dig-holes')
nor that disk usage is "usually much lower from the host's perspective"
(because it's not unusual for a volume to have multiple image files with
lots of diverging data, e.g. revisions).

Hard to say anything there that's true, helpful, and not a novella.