Make error message a bit clearer when killing VM at very early stage
If VM is killed between libvirt_domain.createWithFlags() and
libvirt_domain.resume(), the error message would will be rather
confusing ('NoneType' object has no attribute 'write'). That's because
the connection to QubesDB is done on (first) access to
self.untrusted_qdb and it checks if domain is running at this point.

Note, at this point, domain cannot crash on its own, bacause it isn't
running yet. The only reason for this, is something else killing it. It
could be a manual libvirt call, or (since the previous commit) qvm-kill
call.

Check for this condition to provide a bit better error message.