Draft: lvm2 backend: Disable sync with udev
All of the logical volumes managed by qubesd are ignored by udev, so it
is safe for lvm2 to manage the device nodes in /dev directly.  This
should improve performance.

Marking as draft because udev needs to be made to actually ignore these
volumes, to the extent of not even creating /dev/mapper/* or /dev/VG/LV
symlinks for them.