Refuse to delete an entire volume group
LVM will happily delete the entire contents of a volume group if the
command line requests it, even if it is not empty.  There is typically
a prompt for this, but but Qubes OS disables it with --force.

To prevent accidentally wiping user's systems, check (with an assertion)
that the volume to be deleted is not a bare volume group.

(cherry picked from commit e14956add062cbc68560843350fcab9e5506eca6)