qubesvm: do not prevent suspend if qubes.SuspendPre fails
The service is called to allow VM to prepare for suspend, but it's just
a notification - if the call fails, suspend anyway.
This especially is important for HVM with PCI devices - if the service
fails, userspace might be not prepared for the suspend, but later
suspend by the kernel in many cases is enough for the VM to recover. If
VM isn't suspended (but paused by Xen instead), it almost never survive
resume.

While at it, improve logging of both service failure and actual suspend
failure.

QubesOS/qubes-issues#6066