Use fsync and nocreat during import
It is critical that dd properly flush data to disk, so that imported
data will not be lost in the event of an unclean shutdown.  It is also
critical that dd not try to create the destination file if it does not
exist, as this would try to write a huge amount of data to a tmpfs and
could easily cause out-of-memory if it succeeded.

Fixes QubesOS/qubes-issues#7870.

(cherry picked from commit 87f98660ab6f9ed63cf6fa2addb5c1688310bec1)