Skip test of emulated sound card under PipeWire
PipeWire modules cannot be unloaded, so the test does not work.  It
could be made to work in other ways, but doing so is not worth the
effort.  Both PipeWire and PulseAudio use ALSA to interact with sound
cards, so if one works the other almost certainly will work as well.

Suggested-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>