tests: pass audio data to pw-play via stdin
The .snd extension is apparently not enough to convince pw-play it deals
with raw data, it tries to guess the format by looking for some header
anyway (and ignores explicit format on command line). When it gets the
data via stdin, it doesn't see the file extension at all, and indeed
does use explicit format from the command line then.