Revert "Enable invtsc to let the guest use TSC clocksource, if it wishes so"
Using TSC clocksource requires all VMs (not only those with PCI devices)
to be properly suspended when the system goes to sleep, otherwise the
clock after resume may misbehave. Most of the time, Linux recovers by
detecting TSC is not stable anymore and switching back to clocksource
Xen, but when it doesn't, VM may appear to be frozen.

Kernels before 6.3 chosen clocksource TSC anyway (unless explicit
clocksource=tsc was present on the cmdline), but since 6.3, it prefers
TSC if it's advertised as invariant. Specific Linux commit:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=caea091e48ed9d3951506507abf26e9918d08e35

Qubes OS 4.2 does suspend all VMs properly when the host goes to sleep,
but 4.1 does not, so do not advertise TSC as invariant to avoid
clocksource issues with newer kernels.

This reverts commit 8c8c99c07643c996356f4fc12bbe781a31454f93.