Fix error handling of non-positive/non-zero integers
The property setter helpers for (non-zero) positive integer values
previously raised a ValueError for out of range values, which is
interpreted later as internal error. But in such cases the user supplied
an invalid value so raise a QubesPropertyValueError. The message will
then be displayed by qvm-prefs.

Related to QubesOS/qubes-issues#5005