Drop references to libvirt objects when undefining the domain
Normally it wouldn't matter because dropping reference to
VirDomainWrapper would do. But if there is an exception that happened
during libvirt call (like createWithFlags), the traceback attached to
the exception would keep reference to VirDomainWrapper object,
preventing closing libvirt connection.

During normal runtime it isn't huge problem, as those exceptions are
discarded soon after anyway. But during tests, they are all collected,
and that results in leaking a lot of libvirt connection objects, and as
a consequence open file descriptors.

QubesOS/qubes-issues#