test: extend clipboard tests to various different sizes
Check if copying 64k also works. And then try copying 200kb, both
with default settings (should be refused) and with limit raised. This
test assume already modified behavior on over the limit copy (truncate
to 0 instead of truncating to the limit).
Finally, test copying 300k, which should be rejected regardless of the
limit.

QubesOS/qubes-issues#9296