Use correct internal exception
The exceptions ProtocolError and PermissionDenied are both raised by
qubesd to indicate various problems with requests. However, both cause
clients to print the extremely unhelpful "Got empty response from
qubesd" message.

Use ProtocolError on invalid requests. Use PermissionDenied when the
client doesn't have permission to know something, such as a missing
object, specified argument should be a pool but said pool doesn't exist.

Therefore, a API handler can be arranged as:

- Validation: ProtocolError otherwise
- fire_event_for_permission(): PermissionDenied if unauthorized
- Action

Ideally, validation that may leak existence of a system property should
be done after asking for administrative permission, but then it would
not be possible to pass only safe values to "admin-permission:" event.

If we are already leaking existence of a property, it makes sense to
provide a useful exception class for it.

Strict validation to API method parameters makes it possible to enforce
presence, absence (or both states) for argument and payload.

Don't leave things for chance, require passing explicitly if argument
and payload should be passed or not, and if destination should be
AdminVM.

Passing "**classifiers" means it doesn't check if the argument passed
is a valid classifier or valid value for the classifier. Decided to be
explicit on the allowed classifiers with sane defaults of not setting
the scope at all and only having read capability.

Fixes: https://github.com/QubesOS/qubes-issues/issues/10345