Use ProtocolError on invalid requests
The exceptions ProtocolError and PermissionDenied are both raised by
qubesd to indicate various problems with requests. However, both cause
clients to print the extremely unhelpful "Got empty response from
qubesd" message.

ProtocolError must be used only for client *programming* errors (bugs),
not for problems that the client could not have detected before making
the request.

PermissionDenied must be used only for service authorization denials.

Therefore, a API handler can be arranged as:

- Validation: ProtocolError otherwise
- fire_event_for_permission(): PermissionDenied if unauthorized
- Action

Ideally, validation that may leak existence of a system property should
be done after asking for administrative permission, but then it would
not be possible to pass only safe values to "admin-api:" events.

If we are already leaking existence of a property, it makes sense to
provide a useful exception class for it.

Fixes: https://github.com/QubesOS/qubes-issues/issues/10345