Preload disposables
Is a disposable state, it runs on the background and awaits to intercept
calls made to disposables of the same disposable template, having faster
execution times. Information of the design choices is present in the
DispVM class.

To use them, set the number of allowed preloaded for each disposable
template with the feature "preload-dispvm-max". In case there is not
enough available memory, the maximum won't be preloaded at this
instance, but will retry later on at any relevant event, such as a
preloaded being used or a normal qube being requested while a preload
can be created.

GUI daemon requires a running qube to connect to the GUI agent in the
qube, because of that, auto starting the preload mechanism only happens
after a GUI login.

Preloaded qubes are hidden from GUI applications until the qube itself
is requested to be used. Any GUI application that allows opening
applications on preloaded qubes before they are marked as used is
considered a bug. Applications that autostart on the qube are shown
before the qube has its state interrupted, it is also a bug but that
can't be fixed easily.

For: https://github.com/QubesOS/qubes-issues/issues/1512
For: https://github.com/qubesos/qubes-issues/issues/9918