libvirt-xml: set guivm backend for stubdomain's gui
When using a HVM with a GUI via stubdomain, that GUI agent should
connect to the appropriate guivm. It is dom0 by default, so it works in
a common case, but if somebody uses non-dom0 for GUI, then the parameter
is missing.
Passing through the parameter is already implemented in libvirt, libxl
and QEMU. But setting it in the libvirt xml in the first place was
missing.

Fixes QubesOS/qubes-issues#9385