Remove extra -- from run-tests
It was meant to separate options from package names (end of options
marker), in case any package name would start with a dash. Since package
names are defined a line above, and none of them start with a dash, it
makes no sense to use this marker.
And in fact, it's harmful, since dnf5 doesn't support it in this
position.

Fixes QubesOS/qubes-issues#9797