tests: add simple end-to-end test of remote qrexec calls
Test vm1 -> relay -> vm2-remote call path. Do it on a single host by
having relay service that transform names (source to source-remote and
target-remote to target). For this to work, there need to be matching
pairs of VMs with and without -remote suffix, and appropriate policy for
both outgoing and incoming connections.

For now put the test into 'misc' package, when there will be more
qubes-air related tests, move it elsewhere.

QubesOS/qubes-issues#9015