Move device-related exceptions to the exc module
Admin API, especially the client side depends on having all exceptions
in one place. The exc.py need to be synchronized to core-admin-client
repo. Exceptions from outside this module are translated to a generic
QubesException, which is not perfect.

Specifically, exceptions in wrong place caused DeviceAlreadyAssigned and
DeviceAlreadyAttached being translate to QubesException, which confused
salt module.