vm: Increase connection timeout for qubes.SuspendPost service
The VM may still be resuming from suspend (for example kernel may still
be in progress of re-initializing grant tables), so the default 5s
sometimes is too little. Increase the connection timeout to the full
'suspend_timeout' - quite excessive, but the whole call is under a
timeout too, so it simply makes different type of timeouts handled
uniformly now. Now, the suspend_timeout effectively covers the whole
call, regardless of how much of that time was spent on waiting for the
connection to be established.

Fixes QubesOS/qubes-issues#9942