Minimize unit tests
I tried for a long time to not mock the start() but it was a long
journey and I couldn't make it work. If the qube is not started, it
never calls the qube to be paused, it also can't work to try to unpause
the qube, if it is not being unpaused, it is not marking the qube as
used. Yes, even mocking these events were hard on unit tests as they
were not running. It is being tested what is possible on the unit tests,
but more time or knowledge is necessary to make the full preload suite
be able to be mocked.