Redo netvm setup after unpause
Applying deferred netvm for preloaded disposables is handled separately
to ensure that before a disposable is returned to the user, the
networking is already set up. If the domain-unpaused event of the
NetVMMixin kicks in before the preload is used, it is ignored by the
"is_preload" attribute, if it kicks after, it is ignored by the absent
"deferred-netvm-original" feature.

Fixes: https://github.com/QubesOS/qubes-issues/issues/10173
For: https://github.com/QubesOS/qubes-issues/issues/1512