Clean preloaded disposable remnants on next boot
If the disposable was preloaded, in progress feature would return False,
now it check is qube was a preloaded disposable that completed but is
not running to signal improper shutdown.

When calling cleanup, if the qube is not running, the "_bare_cleanup()"
was skipped on "cleanup()", now this is handled.

There is something that triggers "domain-remove-from-disk" on qubesd
start for unnamed disposables, didn't detect the origin, but it is now
handled by removing the preloaded disposable from the list.

Fixes: https://github.com/qubesos/qubes-issues/issues/10326
For: https://github.com/qubesos/qubes-issues/issues/1512