Remove preloaded disposables on test tearDown
If it is not done on the test, it will be handled by
"qubes/tests/__init__.py", which will attempt to kill the domain. If the
preloaded disposable was still starting, exceptions will be handled by
also attempting to kill the domain. Both methods will trigger the
"_bare_cleanup()", sometimes indirectly via "cleanup()" or
"_auto_cleanup()", but if "_bare_cleanup()" happens on the other call,
not the one that called kill, it will not await and will attempt to
remove the domain from the disk while it is still running (not
completely killed).

For: https://github.com/QubesOS/qubes-core-admin/pull/742
For: https://github.com/QubesOS/qubes-issues/issues/1512
Fixes: https://github.com/qubesos/qubes-issues/issues/10369