tests: adjust for default editor in Whonix 18
Whonix 18 has FeatherPad as default editor, but it doesn't have
application name in the title. Look for just file name, but move that
section at the end, to not intercept any other editor that require
special handling (most editors put the file name in the title).

This change will possibly make test failure less informative - if
another editor shows up, but still has the file name in the title, it
will be handled by this section too. And if it requires different
handling, the test will fail with confusing message. But since many
editors work with the same interaction, it should be a rare issue.