Avoid saving qubes.xml on disposable methods
Saving to file is very useful when you want to deal with sudden stops of
qubesd. Unfortunately, it takes considerable time ~20ms because of
multiple fsyncs, which greatly affects time to use preloaded
disposables. All API calls only call's app.save() once at the end, there
is only one reason it was called multiple times for disposable creation
and usage, to avoid reuse. I tried to reduce the number of saves but
still avoid reuse.

To create a preloaded disposable, there were 4 calls, now, it is 1 to 2
calls. The extra save is to avoid reuse in case preloaded disposable is
not ready yet and it needs to wait, might be unnecessary, leaving
because I am not certain about removing it.

To use a preloaded disposable, there were 3 calls, now, it is 1.

For: https://github.com/qubesos/qubes-issues/issues/10230
For: https://github.com/qubesos/qubes-issues/issues/1512