Skip startup tasks when preload is requested
Waiting for the startup tasks only makes sense when it will be lingering
for some time paused, so the first request is very fast. In case a
request is made while the qube has not finished preloading, skipping the
startup tasks allow faster marking of preloaded disposable as used.

For: https://github.com/qubesos/qubes-issues/issues/10230
For: https://github.com/qubesos/qubes-issues/issues/1512