Skip qmemman task when preload is requested
The qmemman task can take 3 seconds, which is the timeout if transfer speed is
slow. If qube is requested, it is undesirable to wait the remaining time, pause
and unpause. Therefore, skip these stages if preload has been requested.

For: https://github.com/qubesos/qubes-issues/issues/10230
For: https://github.com/qubesos/qubes-issues/issues/1512